Bug fix: strip blank lines from extra args files
If you had any blank lines in your extra args files they were getting
added to the full command line before the optparser call, which meant
they took the position we expect to find the explicit ".files" filename
in. That was leading to an exception when we tried to open the .files
file.